[ADD] Implement initial chat client structure with service and controller. Fix gradle
