+++ build.gradle
@@ -1,27 +1,39 @@
 plugins {
-    id 'org.springframework.boot' version '3.0.0'
-    id 'io.spring.dependency-management' version '1.0.15.RELEASE'
-    id 'java'
+	id 'java'
+	id 'org.springframework.boot' version '3.4.3'
+	id 'io.spring.dependency-management' version '1.1.7'
 }
 
-group = 'com.example'
+group = 'com.springchat'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '23'
+
+java {
+	toolchain {
+		languageVersion = JavaLanguageVersion.of(23)
+	}
+}
 
 repositories {
-    mavenCentral()
+	mavenCentral()
+}
+
+ext {
+	set('springAiVersion', "1.0.0-M6")
 }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter'
-    implementation 'org.springframework.boot:spring-boot-starter-actuator'
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	implementation 'org.springframework.ai:spring-ai-vertex-ai-gemini-spring-boot-starter'
     implementation 'org.springdoc:springdoc-openapi-ui:1.6.14'
-    implementation 'com.google.cloud:google-cloud-aiplatform:2.5.0'
-    implementation 'org.springframework.ai:spring-ai-vertex-ai-gemini-spring-boot-starter'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+}
+
+dependencyManagement {
+	imports {
+		mavenBom "org.springframework.ai:spring-ai-bom:${springAiVersion}"
+	}
 }
 
 tasks.named('test') {
-    useJUnitPlatform()
+	useJUnitPlatform()
 }
