--- src/main/java/com/example/chatclient/config/VertexAIConfig.java
@@ -1,16 +1,0 @@
-package com.example.chatclient.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class VertexAIConfig {
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-
-    // Add any additional beans or configuration properties needed for Google Vertex AI integration
-}
