--- src/main/java/com/example/chatclient/model/ChatMessage.java
@@ -1,42 +1,0 @@
-package com.example.chatclient.model;
-
-import java.time.LocalDateTime;
-
-public class ChatMessage {
-    private String sender;
-    private String content;
-    private LocalDateTime timestamp;
-
-    public ChatMessage() {
-    }
-
-    public ChatMessage(String sender, String content, LocalDateTime timestamp) {
-        this.sender = sender;
-        this.content = content;
-        this.timestamp = timestamp;
-    }
-
-    public String getSender() {
-        return sender;
-    }
-
-    public void setSender(String sender) {
-        this.sender = sender;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public LocalDateTime getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(LocalDateTime timestamp) {
-        this.timestamp = timestamp;
-    }
-}
