+++ src/main/java/com/springchat/chatclient/ChatClientApplication.java
@@ -1,4 +1,4 @@
-package com.example.chatclient;
+package com.springchat.chatclient;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
