+++ src/main/java/com/springchat/chatclient/model/ChatResponse.java
@@ -1,0 +1,3 @@
+package com.springchat.chatclient.model;
+
+public record ChatResponse(String message) {}
