+++ src/main/java/com/springchat/chatclient/service/ChatService.java
@@ -1,0 +1,22 @@
+package com.springchat.chatclient.service;
+
+import com.springchat.chatclient.model.ChatClient;
+import org.springframework.stereotype.Service;
+
+
+@Service
+public class ChatService {
+
+    private final ChatClient chatClient;
+
+    public ChatService(ChatClient.Builder chatClientBuilder) {
+        this.chatClient = chatClientBuilder.build();
+    }
+
+    public String sendMessage(String content) {
+        return this.chatClient.prompt()
+                .user(content)
+                .call()
+                .content();
+    }
+}
